- migrated MSVC project files to MSVC2019 - fix printf format analysis warnings for `%ld`: size_t -> long int

+++ test/file_test.c
@@ -57,7 +57,7 @@ static int compress_file(FL2_CStream *fcs)
         out_size += out_buf.pos;
         out_buf.pos = 0;
     } while (res);
-    fprintf(stdout, "\t%ld -> %ld\n", in_size, out_size);
+    fprintf(stdout, "\t%ld -> %ld\n", (long)in_size, (long)out_size);
     
     return 0;
 
@@ -89,7 +89,7 @@ static int decompress_file(FL2_DStream *fds)
         out_buf.pos = 0;
     } while (res && in_buf.size);
     
-    fprintf(stdout, "\t%ld -> %ld\n", in_size, out_size);
+    fprintf(stdout, "\t%ld -> %ld\n", (long)in_size, (long)out_size);
     
     return 0;
